fix: address PR #217 review feedback
- client: decode api-key expires_at as float64 like every other API
  timestamp — encoding/json errors when unmarshaling a JSON float
  (e.g. 4102444800.0) into int64; convert to whole seconds at the
  provider boundary where the schema exposes Int64. List test now
  exercises float-formatted JSON.
- resource/service_account: add plan-time validators (privilege OneOf
  admin/member/snapshotter/reader; name regex + 1-64 length) and
  UseStateForUnknown on created_at/for_webhook to avoid
  known-after-apply plan noise on updates.
- resource/service_account_api_key: validate description is non-empty
  at plan time.
- deps: add terraform-plugin-framework-validators.

+++ internal/provider/data_source_service_account_api_keys.go
@@ -128,7 +128,7 @@ func (d *serviceAccountAPIKeysDataSource) Read(ctx context.Context, req datasour
 			ID:          types.StringValue(k.ID),
 			Description: types.StringValue(k.Description),
 			CreatedAt:   types.Float64Value(k.CreatedAt),
-			ExpiresAt:   types.Int64Value(k.ExpiresAt),
+			ExpiresAt:   types.Int64Value(int64(k.ExpiresAt)),
 			LastUsedAt:  types.Float64Value(k.LastUsedAt),
 		})
 	}
